+++ Sources/Sentry/SentrySpanDataKey.m
@@ -2,3 +2,8 @@
 
 NSString *const SentrySpanDataKeyFileSize = @"file.size";
 NSString *const SentrySpanDataKeyFilePath = @"file.path";
+
+NSString *const SentrySpanDataKeyAppVitalsStartType = @"app.vitals.start.type";
+NSString *const SentrySpanDataKeyAppVitalsStartValue = @"app.vitals.start.value";
+NSString *const SentrySpanDataKeyAppVitalsStartColdValue = @"app.vitals.start.cold.value";
+NSString *const SentrySpanDataKeyAppVitalsStartWarmValue = @"app.vitals.start.warm.value";