+++ Sources/Swift/Helper/SentryEnabledFeaturesBuilder.swift
@@ -52,6 +52,9 @@ import Foundation
         if options.enableMetrics {
             features.append("metrics")
         }
+        if options.experimental.enableStandaloneAppStartTracing {
+            features.append("standaloneAppStartTracing")
+        }
 
         return features
     }