+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Add standalone app start tracing as an experimental option (#7660), enable it via `options.experimental.enableStandaloneAppStartTracing = true`
+
 ### Fixes
 
 - Fix race conditions in scope observer iteration and propagation context locking (#7897)
@@ -15,6 +19,7 @@
 
 ### Features
 
+- Prevent memory strings in stack overflow crash reports (#7841)
 - Add Set conformance to SentryAttributeValue (#7876)
 
 ## 9.12.1